requirements.txt
# Core
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.11

# Modeling
torch>=2.1
transformers>=4.41
sentence-transformers>=2.7
spacy>=3.7
# optional en_core_web_sm model: python -m spacy download en_core_web_sm

# API
fastapi>=0.110
uvicorn[standard]>=0.29
prometheus-fastapi-instrumentator>=6.1.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-logging>=0.46b0

# Logging to Logstash
python-logstash-async>=2.5

# Testing
pytest>=7.4
httpx>=0.27

# Utilities
pyyaml>=6.0.1
tqdm>=4.66
